pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "docsai-toolkit"
version = "0.1.0"
description = "Docs AI Operating Model utilities (portfolio template)"
requires-python = ">=3.10"
dependencies = ["PyYAML>=6.0"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
